regla de negocios :v

+++ Assets/Scripts/Dominio/Entidades/Cuenta.cs
@@ -11,13 +11,13 @@ public class Cuenta
 {
     private string numero;
     private string numeroInterbancario;
-    private double saldo;
+    private float saldo;
     private Moneda moneda;
     private bool estado;
    
     public Cuenta() { }
 
-    public Cuenta(string numero, string numeroInterbancario, double saldo, Moneda moneda)
+    public Cuenta(string numero, string numeroInterbancario, float saldo, Moneda moneda)
     {
         this.numero = numero;
         this.numeroInterbancario = numeroInterbancario;
@@ -28,10 +28,10 @@ public class Cuenta
 
     public string Numero { get => numero; set => numero = value; }
     public string NumeroInterbancario { get => numeroInterbancario; set => numeroInterbancario = value; }
-    public double Saldo { get => saldo; set => saldo = value; }
+    public float Saldo { get => saldo; set => saldo = value; }
     public bool Estado { get => estado; set => estado = value; }
     public Moneda Moneda { get => moneda; set => moneda = value; }
-    virtual public double calcularComision() 
+    virtual public float calcularComision() 
     {
         return 0.0f;
     }
